removed some unused files from the project.  i will add these in when I am ready to implement them properly, but right now I have done everything in BVH class.
fixed a bug with interpolation and it is working fine with the raw data, however sparse frames are failing.  not sure why this is at this point, but I think it is how I break out the data into sparse form.

worked on the animation player to fix up some timing for interpolation.